12.20.2023
Update Kernel

+++ Kernel/kernel/interface/interface.cpp
@@ -4,7 +4,7 @@
 using namespace Sen::Kernel;
 
 M_EXPORT_API
-unsigned long long version(
+int version(
 
 )
 {
